--- vortex-file/src/read/mod.rs
@@ -1,3 +1,0 @@
-mod record_batch_reader;
-
-pub use record_batch_reader::*;
